+++ docs/conf.py
@@ -73,3 +73,23 @@ html_theme = "pydata_sphinx_theme"
 # html_theme = "sphinx_rtd_theme"
 
 html_static_path = ["_static"]
+
+nbsphinx_prolog = """
+.. raw:: html
+
+    <style>
+        div.rendered_html thead th {
+            text-align: left;
+        }
+        
+        div.rendered_html td {
+            white-space: nowrap;      
+            text-align: left;  
+        }
+        
+        .table {
+            width: 80%;
+            margin-top: 50px;
+        }
+    </style>
+"""